\name{Lform}
\alias{Lform}
\title{Hypothesis matrix generated by expressions}
\usage{
Lform(fit, form, data = getData(fit))
}
\arguments{

\item{fit}{a fitted model with a \code{\link{getFix}} method.}

\item{form}{formulas (expressions) to evaluate (see \emph{Details}).}

\item{data}{the data frame in which expressions are evaluated.}

}
\value{
hypothesis matrix
}
\description{
Creates an L matrix using expressions evaluated in \code{data} for each column of
the L matrix
}
\details{
If \code{Lform} is called with only a \code{fit} argument, it outputs code
consisting of an expression that would, if used as the \code{form} argument
to \code{Lform} would generate the full model matrix for the linear model.

If \code{Lform} is called with two or three arguments, it generates a
hypothesis matrix by evaluating the expressions in \code{form} in the
environment \code{data}. The function \code{M} is designed to facilitate the
generation of blocks of the hypothesis matrix corresponding to main effects
or interaction effects of factors.
}


\examples{
mod <- lm( income ~ (education + I(education^2) )* type, Prestige)
summary(mod)

# estimate the marginal value of an extra year of education for a
# range of years for each type

years.type <- expand.grid( education = seq(6,18,2), type = levels(Prestige$type))
Lf <- Lform( mod,
   list( 0, 1, 2*education, 0, 0, type =="prof", type =="wc",
      2*education*(type =="prof"), 2*education*(type =="wc")),
   years.type)
Lf
ww <- wald( mod, Lf)
ww
ytderiv <- as.data.frame( ww, se = 2)
head( ytderiv )
if (require("lattice")){
   xyplot(coef ~ education, ytderiv, groups = type, type = 'l',
           auto.key = list(columns = 3, lines = TRUE, points = FALSE))
   }
}

\seealso{
\code{\link{wald}}
}
